move button component and declare props interface in same file
