+++ wbooks/src/app/components/Button/index.tsx
@@ -1,8 +1,14 @@
 import React from 'react';
 import { Text, TouchableOpacity } from 'react-native';
 
-import { ButtonProps } from '../../../interfaces/button';
 import styles from './styles';
+
+interface ButtonProps {
+  text: string;
+  fill?: boolean;
+  onPress?: any;
+  disabled?: boolean;
+}
 
 const Button = ({ text, fill, onPress, disabled }: ButtonProps) => (
   <TouchableOpacity
