--- wbooks/src/app/interfaces/button.ts
@@ -1,6 +1,0 @@
-export interface ButtonProps {
-  text: string;
-  fill?: boolean;
-  onPress?: any;
-  disabled?: boolean;
-}
